refactor: split advanced.rs into pipeline/ subdirectory (PR-4c)
Decomposes the 2,027-line src/memory_core/storage/sqlite/advanced.rs into
a sqlite/pipeline/ subdirectory of phase modules:

- pipeline/retrieval.rs   -- Phase 1 (vector) + Phase 2 (FTS5)
- pipeline/rerank.rs      -- Phase 2.5 cross-encoder reranking
- pipeline/fusion.rs      -- Phase 3 RRF + dual-match boost + cross-encoder blend
- pipeline/scoring.rs     -- Phase 4 score refinement + keyword conversion
- pipeline/enrichment.rs  -- Phase 5 graph neighbors + Phase 5b entity tags
- pipeline/abstention.rs  -- Phase 6 dedup + abstention + hot-cache merge helpers
- pipeline/decomp.rs      -- run_single_query_pipeline orchestrator

advanced.rs residual (632 lines) keeps `impl FtsSearcher`, `impl AdvancedSearcher`
(orchestrating calls into pipeline::*), and the existing `#[cfg(test)] mod tests`
submodule. Pure code move -- no behavior changes.

The 632-line residual exceeds the 500-line target documented in
docs/specs/module-decomposition.md §4: 381 lines is the `advanced_search`
orchestration body (which pipeline/ exists to support and which can't be split
further without functional changes), and 203 lines is the test submodule (kept
co-located per the spec). Pipeline submodules are all under 500 lines
(largest: enrichment.rs at 438).

Closes Phase 4 of the Substrate Campaign per
docs/specs/execution-roadmap.md §PR-4c.

Quality gate note: cargo/prek/bench could not run locally (sandbox blocks
ort-sys binary downloads). CI will exercise prek + bench --gate.
